+++ src/main/java/command/CommandType.java
@@ -1,3 +1,4 @@
+package command;
 public enum CommandType {
     LIST,
     MARK,
