+++ src/main/java/command/DeleteCommand.java
@@ -1,4 +1,13 @@
+package command;
 import java.io.IOException;
+
+import task.TaskList;
+
+import exception.ScheduloException;
+
+import util.Storage;
+
+import util.Ui;
 
 public class DeleteCommand extends Command {
     private int index;
