+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import task.TaskList;
+
+import util.Storage;
+
+import util.Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
