+++ src/main/java/command/MarkCommand.java
@@ -1,15 +1,24 @@
+package command;
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+import task.TaskList;
+
+import exception.ScheduloException;
+
+import util.Storage;
+
+import util.Ui;
+
+public class MarkCommand extends Command {
     private int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws ScheduloException, IOException {
-        tasks.unmark(index - 1);
+        tasks.mark(index - 1);
         storage.save(tasks);
     }
 }
