+++ src/main/java/exception/ScheduloException.java
@@ -1,3 +1,4 @@
+package exception;
 public class ScheduloException extends Exception {
 
     public ScheduloException(String message) {
