+++ src/main/java/task/Event.java
@@ -1,3 +1,4 @@
+package task;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDateTime;
