+++ src/main/java/util/Ui.java
@@ -1,3 +1,4 @@
+package util;
 import java.util.Scanner;
 
 public class Ui {
